Updated Maze game to use PyGame when a RasperberryPi is not connected.
