+++ games/__init__.py
@@ -1,3 +1,3 @@
-from .missile_defense import MissileDefense
-from .pong import Pong
-from .maze import Maze
+# from .missile_defense import MissileDefense
+# from .pong import Pong
+# from .maze import Maze
