+++ requirements.txt
@@ -1,5 +1,6 @@
-smbus2
-RPi.GPIO
-gpiozero
-cachetools
-scipy
+cachetools==3.0.0
+gpiozero==1.4.1
+numpy==1.15.4
+RPi.GPIO==0.6.5
+scipy==1.1.0
+smbus2==0.2.1
